requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Pydantic for data validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.0
psycopg[binary]>=3.1.8
asyncpg>=0.29.0

# Authentication (versiones específicas compatibles)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6

# Environment and configuration
python-dotenv>=1.0.0

# Production server
gunicorn>=21.2.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
aiosqlite>=0.19.0

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0